+++ Android_AccMvvm/app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
             android:name=".streaming.StreamingActivity"
             android:exported="false" />
         <activity
-            android:name=".StartActivity"
+            android:name=".start.StartActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
