+++ Android_AccMvvm/app/src/main/res/layout/activity_start.xml
@@ -4,16 +4,13 @@
     xmlns:tools="http://schemas.android.com/tools">
     
     <data>
-        <variable
-            name="viewModel"
-            type="com.example.ivcs_android.StartViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#EAEAEA"
-        tools:context=".StartActivity">
+        tools:context=".start.StartActivity">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -30,7 +27,6 @@
                 android:layout_height="wrap_content"
                 android:layout_margin="10dp"
                 android:background="@drawable/radius_rec"
-                android:onClick="@{ () -> viewModel.btGoToStreamingClicked() }"
                 android:text="Streaming"
                 android:textColor="#000000" />
 
@@ -40,7 +36,6 @@
                 android:layout_height="wrap_content"
                 android:layout_margin="10dp"
                 android:background="@drawable/radius_rec"
-                android:onClick="@{()->viewModel.btGoToAnalysisClicked()}"
                 android:text="Analysys"
                 android:textColor="#000000" />
         </LinearLayout>
